First prototype for API - nManager
