+++ src/main/java/com/redhat/patriot/network_simulator/example/Main.java
@@ -5,7 +5,7 @@ import com.github.dockerjava.core.DefaultDockerClientConfig;
 import com.github.dockerjava.core.DockerClientBuilder;
 import com.redhat.patriot.network_simulator.example.args.CommandLineParser;
 import com.redhat.patriot.network_simulator.example.cleanup.Cleaner;
-import com.redhat.patriot.network_simulator.example.manager.DockerManager;
+import com.redhat.patriot.network_simulator.example.manager.DockerController;
 import org.kohsuke.args4j.CmdLineException;
 import org.kohsuke.args4j.CmdLineParser;
 import org.slf4j.Logger;
@@ -18,7 +18,7 @@ public class Main {
     private static final Logger LOGGER = LoggerFactory.getLogger(Main.class);
 
     public static void main(String[] args) throws InterruptedException {
-        DockerManager dockerManager = new DockerManager();
+        DockerController dockerController = new DockerController();
         CommandLineParser cmdArgs = new CommandLineParser();
         CmdLineParser parser = new CmdLineParser(cmdArgs);
 
@@ -37,7 +37,7 @@ public class Main {
                     Arrays.asList("comm_client", "comm_server", "router"));
         } else {
             LOGGER.info("Generating enviropment");
-            dockerManager.genererateEnviroment();
+            dockerController.genererateEnviroment();
         }
