+++ src/main/java/com/redhat/patriot/network_simulator/example/manager/Manager.java
@@ -4,10 +4,11 @@ import com.redhat.patriot.network_simulator.example.container.Container;
 import com.redhat.patriot.network_simulator.example.network.Network;
 
 import java.util.List;
+import java.util.Set;
 
 public interface Manager {
-    Container createContainerResponse();
-    Network createNetwork();
+    Container createContainerResponse(String name, String tag);
+    Network createNetwork(String name, String subnet);
     List<Container> listContainers();
 
 }
